RaspiPicoファームウェアの更新(v1.2) - MuSICAの楽曲データの再生に対応した（KINROU5.DRVを使用）。ただしMGSDRVとは排他使用となる。●SW押下しながら電源ONするとMuSICA再生モードになる。 - 16ms周期で割り読み処理を呼び出しす処理で、16ms以上時間がかかったとき、次の周期の時間を短くして曲の長さの帳尻を合わせるようにした（限度あり） - HopStepZ:メインROM 002Dh を参照された場合、02h(MSX2+)を返すようにした。 - リファクタリング

+++ RasPiPico/rp2040/mgspico/CUTimeCount.cpp
@@ -28,6 +28,6 @@ uint64_t CUTimeCount::GetTime()
 
 uint64_t CUTimeCount::getCount()	// マイクロ秒単位
 {
-	uint64_t usec = GetTimerCounterMS() * 1000L; // msからus単位に。
+	uint64_t usec = (uint64_t)GetTimerCounterMS() * 1000L; // msからus単位に。
 	return usec;
 }
